+++ packages/components/atoms/f-popover/stories/Popover.stories.js
@@ -1,20 +1,32 @@
-// Uncomment the import below to add prop controls to your Story (and add `withKnobs` to the decorators array)
-// import {
-//     withKnobs, select, boolean
-// } from '@storybook/addon-knobs';
 import { withA11y } from '@storybook/addon-a11y';
-import VPopover from '../src/components/Popover.vue';
+import Popover from '../src/components/Popover.vue';
 
 export default {
     title: 'Components/Atoms',
     decorators: [withA11y]
 };
 
-export const VPopoverComponent = () => ({
-    components: { VPopover },
+export const PopoverComponent = () => ({
+    components: { Popover },
     props: {
     },
-    template: `<v-popover />`
+    template: `
+        <div style="position: relative; padding: 16px;">
+            Some text
+            <popover 
+                style="position: absolute;
+                        top: 100%;
+                        width: 300px;
+                        opacity: 1;
+                        z-index: 99999;
+                        left: 0;">
+                <ul>
+                    <li>List item 1</li>
+                    <li>List item 2</li>
+                    <li>List item 3</li>
+                </ul>
+            </popover>
+        </div>`
 });
 
-VPopoverComponent.storyName = 'f-popover';
+PopoverComponent.storyName = 'f-popover';