+++ packages/components/molecules/f-mega-modal/test-utils/component-objects/f-mega-modal.component.js
@@ -2,6 +2,8 @@ const Page = require('@justeat/f-wdio-utils/src/page.object');
 
 module.exports = class MegaModal extends Page {
     get component () { return $('[data-test-id="mega-modal-component"]'); }
+    get megaModalTitle () { return $('[data-test-id="mega-modal-title"]'); }
+    get megaModalContent () { return $('[data-test-id="mega-modal-content"]'); }
 
     open () {
         super.openComponent('molecule', 'mega-modal-component');
@@ -10,4 +12,17 @@ module.exports = class MegaModal extends Page {
     waitForComponent () {
         super.waitForComponent(this.component);
     }
+
+    isComponentDisplayed () {
+        return this.component.isDisplayed();
+    }
+
+    isTitleDisplayed () {
+        return this.megaModalTitle.isDisplayed();
+    }
+
+    isContentDisplayed () {
+        return this.megaModalTitle.isDisplayed();
+    }
+
 };