+++ packages/components/molecules/f-searchbox/test/specs/component/f-searchbox.component.spec.js
@@ -7,7 +7,7 @@ describe('f-searchbox component tests', () => {
         searchbox.waitForComponent();
     });
 
-    it.skip('should display the f-searchbox component', () => {
+    it('should display the f-searchbox component', () => {
         // Assert
         expect(searchbox.isComponentDisplayed()).toBe(true);
     });