+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -107,7 +107,13 @@ import GuestBlock from './Guest.vue';
 import UserNote from './UserNote.vue';
 import ErrorPage from './Error.vue';
 
-import { CHECKOUT_METHOD_DELIVERY, TENANT_MAP, VALIDATIONS } from '../constants';
+import {
+    CHECKOUT_METHOD_DELIVERY,
+    TENANT_MAP,
+    VALIDATIONS,
+    VUEX_CHECKOUT_ANALYTICS_MODULE,
+    VUEX_CHECKOUT_MODULE
+} from '../constants';
 import checkoutValidationsMixin from '../mixins/validations.mixin';
 import EventNames from '../event-names';
 import tenantConfigs from '../tenants';
@@ -232,7 +238,7 @@ export default {
     },
 
     computed: {
-        ...mapState('checkout', [
+        ...mapState(VUEX_CHECKOUT_MODULE, [
             'address',
             'customer',
             'id',
@@ -293,7 +299,7 @@ export default {
     },
 
     methods: {
-        ...mapActions('checkout', [
+        ...mapActions(VUEX_CHECKOUT_MODULE, [
             'createGuestUser',
             'getAvailableFulfilment',
             'getAddress',
@@ -307,7 +313,7 @@ export default {
             'getGeoLocation'
         ]),
 
-        ...mapActions('analytics', [
+        ...mapActions(VUEX_CHECKOUT_ANALYTICS_MODULE, [
             'trackInitialLoad',
             'trackFormInteraction'
         ]),
@@ -546,7 +552,7 @@ export default {
                 this.$emit(EventNames.CheckoutAddressGetSuccess);
             } catch (thrownErrors) {
                 this.$emit(EventNames.CheckoutAddressGetFailure, thrownErrors);
-                this.hasCheckoutLoadedSuccessfully = false;
+                this.$logger.logWarn('Get checkout address failure', this.$store, { thrownErrors });
             }
         },