+++ packages/components/organisms/f-checkout/src/store/checkoutAnalytics.module.js
@@ -1,5 +1,5 @@
-import Trak from '@justeat/f-trak';
 import { mapAnalyticsName, mapAnalyticsNames } from '../services/mapper';
+import { VUEX_CHECKOUT_MODULE } from '../constants';
 
 import {
     UPDATE_CHANGED_FIELD,
@@ -51,23 +51,23 @@ export default {
          * Pushes initial state of checkout to the dataLayer.
          */
         trackInitialLoad ({ rootState, dispatch }) {
-            window.dataLayer = window.dataLayer || [];
+            if (typeof (window) === 'undefined') {
+                return;
+            }
 
-            const pageName = rootState.checkout.isLoggedIn ? 'Overview' : 'Guest';
+            const pageName = rootState[VUEX_CHECKOUT_MODULE].isLoggedIn ? 'Overview' : 'Guest';
 
-            Trak.event({
-                custom: {
-                    checkout: {
-                        step: 1
-                    },
-                    basket: rootState.checkout.basket,
-                    restaurant: {
-                        id: rootState.checkout.restaurantId
-                    },
-                    pageData: {
-                        name: `Checkout 1 ${pageName}`,
-                        group: 'Checkout'
-                    }
+            window.dataLayer.push({
+                checkout: {
+                    step: 1
+                },
+                basket: rootState[VUEX_CHECKOUT_MODULE].basket,
+                restaurant: {
+                    id: rootState[VUEX_CHECKOUT_MODULE].restaurantId
+                },
+                pageData: {
+                    name: `Checkout 1 ${pageName}`,
+                    group: 'Checkout'
                 }
             });
 
@@ -78,20 +78,18 @@ export default {
          * Pushes `form` event to the dataLayer with correct data
          */
         trackFormInteraction ({ state, rootState }, { action, error }) {
-            const formName = rootState.checkout.isLoggedIn ? 'checkout' : 'checkout_guest';
+            const formName = rootState[VUEX_CHECKOUT_MODULE].isLoggedIn ? 'checkout' : 'checkout_guest';
 
             const mappedError = error ? mapAnalyticsNames(error) : null;
 
-            Trak.event({
+            window.dataLayer.push({
                 event: 'Form',
-                custom: {
-                    form: {
-                        name: formName,
-                        action,
-                        error: mappedError,
-                        autofill: state.autofill,
-                        changes: state.changedFields.sort()
-                    }
+                form: {
+                    name: formName,
+                    action,
+                    error: mappedError,
+                    autofill: state.autofill,
+                    changes: state.changedFields
                 }
             });
         }
@@ -102,6 +100,8 @@ export default {
             if (!state.changedFields.includes(field)) {
                 state.changedFields.push(field);
             }
+
+            state.changedFields.sort();
         },
 
         [UPDATE_AUTOFILL]: (state, autofill) => {