+++ packages/components/organisms/f-header/src/tenants/en-NZ.js
@@ -4,6 +4,7 @@ export default {
     openMenuText: 'Open Menu',
     loginNoScriptLinkText: 'Account',
     skipToMainContentText: 'Skip to main content',
+    navTitle: 'Main menu',
 
     navLinks: {
         accountInfo: {