+++ packages/components/molecules/f-mega-modal/CHANGELOG.md
@@ -3,9 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-
-Latest (add to next release)
+v0.5.0
 ------------------------------
+*March 3, 2021*
+
+### Added
+- Added component test file, component object file and tests
+- Added `Data-test-id` to elements within the `MegaModal.stories.js`
+
 *February 25, 2021*
 
 ### Changed